refactor: simplify sorting logic
